phone _agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Self-hosted SIP AI voice agent using Asterisk AGI + Google APIs"
requires-python = ">=3.11"
dependencies = [
  "google-auth>=2.29.0",
  "google-cloud-speech>=2.26.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
